Organize code on ready page
